Modularized /auth route into middleware and router; tests pass but still need to work on error handling (global error handler) and bearer token verificationmiddleware
